% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_lsn_to_temp.R
\name{copy_lsn_to_temp}
\alias{copy_lsn_to_temp}
\title{Copy LSN to temporary directory}
\usage{
copy_lsn_to_temp()
}
\value{
A copy of MiddleFork04.ssn residing in R's temporary directory
}
\description{
Copies the LSN directory MiddleFork04.ssn to R's temporary
directory so the examples in SSN2 do not write to the local
library or any other places.
}
\details{
Copies the LSN directory MiddleFork04.ssn to R's temporary directory
}
\examples{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
# getwd()
# setwd(tempdir())
# getwd()
# if unix-alike, list temporary directory contents using: system('ls')
# if windows, list temporary directory contents using: shell('dir')
}
