% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf04p}
\alias{mf04p}
\title{Imported SSN object from the MiddleFork04.ssn data folder}
\format{
An object of class \code{SSN} of length 4.
}
\usage{
mf04p
}
\description{
The MiddleFork04.ssn data folder contains the spatial, attribute,
and topological information needed to construct a spatial stream
network object using the SSN2 package. \code{mf04p} was created
using \code{\link[=ssn_import]{ssn_import()}}.
}
\seealso{
\link{MiddleFork04.ssn} for details about the contents of \code{mf04p}.
\code{\link[=ssn_import]{ssn_import()}} to convert a .ssn object to an \code{SSN} object in R.
\link{ssn_create_distmat} for details about the distance matrix file structure.
}
\keyword{datasets}
