% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.SSN2}
\alias{nobs.SSN2}
\alias{nobs.ssn_lm}
\alias{nobs.ssn_glm}
\title{Extract the number of observations}
\usage{
\method{nobs}{ssn_lm}(object, ...)

\method{nobs}{ssn_glm}(object, ...)
}
\arguments{
\item{object}{A fitted model object from \code{\link[=ssn_lm]{ssn_lm()}} or \code{\link[=ssn_glm]{ssn_glm()}}.}

\item{...}{Other arguments. Not used (needed for generic consistency).}
}
\value{
The number of observations used for modeling (i.e., \code{object$n}).
}
\description{
Find the number of observations used for model fitting. This
exists so that generic R functions that call \code{nobs()} (e.g.,
\code{stats::nobs()}) work on \code{SSN2} fitted model objects.
Could alternatively rename object$n as object$nobs and rely on stats::nobs.default.
}
\examples{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)

ssn_mod <- ssn_lm(
  formula = Summer_mn ~ ELEV_DEM,
  ssn.object = mf04p,
  tailup_type = "exponential",
  additive = "afvArea"
)
nobs(ssn_mod)
}
