% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SSN.R
\name{print.SSN}
\alias{print.SSN}
\title{Print SSN object}
\usage{
\method{print}{SSN}(x, ...)
}
\arguments{
\item{x}{An SSN object.}

\item{...}{Other arguments. Not used (needed for generic consistency).}
}
\value{
Print summary to console
}
\description{
Print information about the data found in an SSN object.
}
