% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate_ssn.R, R/print.R, R/print_rf.R
\name{print.ssn_decorrelate}
\alias{print.ssn_decorrelate}
\alias{print.SSN2}
\alias{print.ssn_lm}
\alias{print.ssn_glm}
\alias{print.summary.ssn_lm}
\alias{print.summary.ssn_glm}
\alias{print.anova.ssn_lm}
\alias{print.anova.ssn_glm}
\alias{print.ssn_lmRF}
\alias{print.summary.ssn_lmRF}
\title{Print values}
\usage{
\method{print}{ssn_decorrelate}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{ssn_lm}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{ssn_glm}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{summary.ssn_lm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{summary.ssn_glm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{anova.ssn_lm}(
  x,
  digits = max(getOption("digits") - 2L, 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{anova.ssn_glm}(
  x,
  digits = max(getOption("digits") - 2L, 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{ssn_lmRF}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{summary.ssn_lmRF}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A fitted model object from \code{\link[=ssn_lm]{ssn_lm()}}, \code{\link[=ssn_glm]{ssn_glm()}},
\code{\link[=ssn_decorrelate]{ssn_decorrelate()}}, or \code{\link[=ssn_lmRF]{ssn_lmRF()}}, or output from \code{summary(x)}
or \code{anova(x)}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Other arguments passed to or from other methods.}

\item{signif.stars}{Logical. If \code{TRUE}, significance stars are printed for each coefficient}
}
\value{
Printed fitted model objects and summaries with formatting.
}
\description{
Print fitted model objects and summaries.
}
\examples{
\donttest{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)
fit <- ssn_lm(
  Summer_mn ~ ELEV_DEM, mf04p, tailup_type = "exponential",
  additive = "afvArea"
)
print(fit)
}
}
