% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{augment}
\alias{AICc}
\alias{AUROC}
\alias{conditional}
\alias{covmatrix}
\alias{dispersion_initial}
\alias{dispersion_params}
\alias{glances}
\alias{kcv}
\alias{loocv}
\alias{pseudoR2}
\alias{randcov_initial}
\alias{randcov_params}
\alias{satterthwaite}
\alias{varcomp}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{spmodel}{\code{\link[spmodel]{AICc}}, \code{\link[spmodel]{AUROC}}, \code{\link[spmodel]{conditional}}, \code{\link[spmodel]{covmatrix}}, \code{\link[spmodel]{dispersion_initial}}, \code{\link[spmodel]{dispersion_params}}, \code{\link[spmodel]{glances}}, \code{\link[spmodel]{kcv}}, \code{\link[spmodel]{loocv}}, \code{\link[spmodel]{pseudoR2}}, \code{\link[spmodel]{randcov_initial}}, \code{\link[spmodel]{randcov_params}}, \code{\link[spmodel]{satterthwaite}}, \code{\link[spmodel]{varcomp}}}
}}

