% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{satterthwaite.SSN2}
\alias{satterthwaite.SSN2}
\alias{satterthwaite.ssn_lm}
\alias{satterthwaite.ssn_glm}
\title{Compute Satterthwaite denominator degrees of freedom}
\usage{
\method{satterthwaite}{ssn_lm}(object, method, ...)

\method{satterthwaite}{ssn_glm}(object, ...)
}
\arguments{
\item{object}{A fitted model object from \code{\link[=ssn_lm]{ssn_lm()}} fit without \code{local}.
Not supported for \code{\link[=ssn_glm]{ssn_glm()}} objects.}

\item{method}{The method by which to compute gradients. Must be supplied
as \code{"numeric"} for numerical differentiation, the only supported method.}

\item{...}{Other arguments. Not used (needed for generic consistency).}
}
\value{
A named numeric vector of Satterthwaite degrees of freedom for each
fixed effect.
}
\description{
Compute Satterthwaite denominator degrees of freedom
\eqn{t}-based (rather than asymptotic \eqn{z}-based)
fixed effect inference in small samples.
}
\details{
Satterthwaite degrees of freedom are generally more appropriate than
asymptotic degrees of freedom for small samples. They can be computationally costly
for sample sizes exceeding 500; however, for sample sizes this large, they Satterthwaite
and asymptotic degrees of freedom should yield very similar inferences.
}
\examples{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)

ssn_mod <- ssn_lm(
  formula = Summer_mn ~ ELEV_DEM,
  ssn.object = mf04p,
  tailup_type = "exponential",
  additive = "afvArea"
)
satterthwaite(ssn_mod, method = "numeric")

}
\references{
Rencher, Alvin C. and Schaalje, G. Bruce (2008). Linear Models in
Statistics, Second Edition. John Wiley & Sons.
}
\seealso{
\code{\link[=ssn_lm]{ssn_lm()}} \code{\link[=anova.SSN2]{anova.SSN2()}}
}
