% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSN2.R
\docType{package}
\name{SSN2-package}
\alias{SSN2}
\alias{SSN2-package}
\title{SSN2: Spatial Modeling on Stream Networks}
\description{
Fit, summarize, and predict for spatial linear and generalized linear models on stream networks. Stream covariance models use moving average constructions based on in-stream distance. Parameters are estimated using likelihood methods. Additional modeling features include Euclidean covariance, anisotropy, non-spatial random effects, partition factors, and big data methods. Model-fit statistics are used to summarize, visualize, and compare models, and predictions at unobserved locations are readily obtainable. For additional details, see Ver Hoef and Peterson (2010) \doi{10.1198/jasa.2009.ap08248} and Dumelle et al. (2024) \doi{10.21105/joss.06389}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usepa.github.io/SSN2/}
  \item Report bugs at \url{https://github.com/USEPA/SSN2/issues}
}

}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Authors:
\itemize{
  \item Jay M. Ver Hoef \email{jay.verhoef@noaa.gov}
  \item Erin Peterson
}

Other contributors:
\itemize{
  \item Alan Pearse [contributor]
  \item Dan Isaak [contributor]
  \item Ryan A. Hill [contributor]
  \item Marc Weber [contributor]
}

}
\keyword{internal}
