% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate_ssn.R
\name{ssn_decorrelate}
\alias{ssn_decorrelate}
\alias{tidy.ssn_decorrelate}
\title{Apply the Spatial Decorrelation Transformation for Machine Learning Models}
\usage{
ssn_decorrelate(
  formula,
  ssn.object,
  tailup_type = "none",
  taildown_type = "none",
  euclid_type = "none",
  nugget_type = "nugget",
  tailup_params,
  taildown_params,
  euclid_params,
  nugget_params,
  additive,
  algorithm = "ranger",
  statistic = "RMSPE",
  training,
  evaluate_test,
  anisotropy = FALSE,
  random,
  randcov_params,
  partition_factor,
  ordering,
  local,
  grid,
  dense_grid,
  ...
)

\method{tidy}{ssn_decorrelate}(x, ...)
}
\arguments{
\item{formula}{A two-sided linear formula describing the fixed effect structure
of the model, with the response to the left of the \code{~} operator and
the terms on the right, separated by \code{+} operators. \code{.} on the
right-hand side represents every variable in \code{data} except the geometry column.}

\item{ssn.object}{A spatial stream network object with class \code{SSN}.}

\item{tailup_type, taildown_type, euclid_type, nugget_type}{Covariance types
specified as in \code{\link[=ssn_lm]{ssn_lm()}}.}

\item{tailup_params, taildown_params, euclid_params, nugget_params}{Known
covariance parameter objects to be used in the spatial decorrelation
transformation. See \code{\link[=tailup_params]{tailup_params()}}, \code{\link[=taildown_params]{taildown_params()}},
\code{\link[=euclid_params]{euclid_params()}}, and \code{\link[=nugget_params]{nugget_params()}}.}

\item{additive}{Additive-function variable for the tail-up covariance, as in \code{\link[=ssn_lm]{ssn_lm()}}.}

\item{algorithm}{The machine learning algorithm applied. Available options
include \code{"ranger"}, \code{"randomForest"}, and \code{"xgboost"}.
\code{"ranger"} specifies a random forest via \code{\link[ranger:ranger]{ranger::ranger()}}.
\code{"randomForest"} specifies a random forest via \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.
\code{"xgboost"} specifies a boosted decision tree ensemble via \code{\link[xgboost:xgboost]{xgboost::xgboost()}}.}

\item{statistic}{The statistic used to evaluate fit in the test data. Available options
include \code{"bias"} (mean bias), \code{"MSPE"} (mean-squared-prediction error),
\code{"RMSPE"} (root-mean-squared-prediction error)
and \code{"cor2"} (the predictive R-squared; i.e., the
squared correlation between observations and predictions).}

\item{training}{A list controlling how the training and test data are assigned
when evaluating test data performance.
The following arguments detail this process:
\itemize{
\item \code{method}: The method used to evaluate test data performance.
\code{"split"} will split \code{data} up into distinct training and test sets
proportionally based on \code{p}. \code{"cv"} will split \code{data} up
via k-fold cross validation based on \code{folds}, the number of folds.
\item \code{p}: The proportion (a numeric vector between zero and one) of observations in \code{data} that should
be assigned to the training data. The default is 0.8, which means that
80\% of the observations are assigned to the training data and 20\% to the
test data. Ignored if \code{training_index} or \code{test_index} are provided.
\item \code{replicate}: The number of times to replicate \code{"split"} with different random training and test assignments.
\item \code{folds}: The number of folds to use in cross-validation. Requires \code{method = "cv"}. Ignored if \code{folds_index} is specified. The default is 5, matching the 80/20 default split above.
\item \code{training_index}: A numeric vector that specifies which rows (i.e., indices)
of \code{data} should be assigned to the training data. If omitted, defaults
to the rows which are not already included in \code{test_index}.
\item \code{test_index}: A numeric vector that specifies which rows (i.e., indices)
of \code{data} should be assigned to the test data. If omitted, defaults
to the rows which are not already included in \code{training_index}.
\item \code{folds_index}: A numeric vector that specifies which rows
of \code{data} are associated with each cross-validation fold. Requires
\code{method = "cv"}.
}
If omitted, \code{training} is transformed into
\code{list(method = "split", p = 0.8, replicate = 1)}.}

\item{evaluate_test}{A logical indicating whether a grid should be constructed
and evaluated when spatial decorrelation parameters are known (i.e.,
\code{*_params} are all specified, and, if random effects are included, \code{randcov_params} is specified).
If \code{TRUE}, constructs and evaluates the grid after assigning observations to
test and training data sets. If any parameters (spatial or random effects) are estimated via a grid search,
\code{evaluate_test} is set to \code{TRUE}.}

\item{anisotropy}{A logical indicating whether Euclidean (geometric) anisotropy should
be modeled. Not required if \code{euclid_params} is provided with a nonzero
\code{rotate} or a \code{scale} less than one. When \code{anisotropy} is
\code{TRUE}, computational times can significantly increase. The default
is \code{FALSE}.}

\item{random}{A one-sided linear formula describing the random effect structure
of the model. Terms are specified to the right of the \code{~ operator}.
Each term has the structure \code{x1 + ... + xn | g1/.../gm}, where \code{x1 + ... + xn}
specifies the model for the random effects and \code{g1/.../gm} is the grouping
structure. Separate terms are separated by \code{+} and must generally
be wrapped in parentheses. Random intercepts are added to each model
implicitly when at least one other variable is defined.
If a random intercept is not desired, this must be explicitly
defined (e.g., \code{x1 + ... + xn - 1 | g1/.../gm}). If only a random intercept
is desired for a grouping structure, the random intercept must be specified
as \code{1 | g1/.../gm}. Note that \code{g1/.../gm} is shorthand for \code{(1 | g1/.../gm)}.
If only random intercepts are desired and the shorthand notation is used,
parentheses can be omitted.}

\item{randcov_params}{An optional random effect covariance object
to be used in the spatial decorrelation transformation. See \code{\link[spmodel:randcov_params]{spmodel::randcov_params()}}.}

\item{partition_factor}{A one-sided linear formula with a single term
specifying the partition factor. The partition factor assumes observations
from different levels of the partition factor are uncorrelated.}

\item{ordering}{The data ordering applied: Available options
include \code{"pid"}, \code{"grts"}, \code{"maxmin"}, \code{"middleout"},
\code{"outsidein"}, \code{"coordinate"}, \code{"random"}, and \code{"none"}.
\code{"pid"} applies ordering by point ID from the network geometry.
\code{"grts"} applies ordering using a spatially balanced GRTS sample via \code{spsurvey::grts()}.
\code{"maxmin"} applies maximum minimum distance ordering via \code{GPvecchia::order_maxmin_exact()}.
\code{"middleout"} applies middle out ordering via \code{GPvecchia::order_middleout()}.
\code{"outsidein"} applies middle out ordering via \code{GPvecchia::order_outsidein()}.
\code{"coordinate"} applies middle out ordering via \code{GPvecchia::order_coordinate(..., coordinate = c(1, 2))},
which orders from bottom-left to top-right of the spatial domain.
\code{"random"} applies a completely random ordering.
\code{"none"} applies no random ordering.
The default is \code{"pid"}.}

\item{local}{An optional logical or list controlling the big data approximation.
If omitted, \code{local} is set
to \code{TRUE} or \code{FALSE} based on the sample size (the number of
non-missing observations in \code{data}) -- if the sample size exceeds 5,000,
\code{local} is set to \code{TRUE}. Otherwise it is set to \code{FALSE}.
If \code{local} is \code{FALSE}, no big data approximation
is implemented. If a list is provided, the following arguments detail the big
data approximation:
\itemize{
\item \code{method}: The big data approximation method. If \code{method = "all"},
all observations are used and \code{size} is ignored.
If \code{method = "covariance"}, the \code{size} data observations
with the highest covariance with the observation requiring prediction are used.
The default is \code{"covariance"}.
\item \code{size}: The number of data observations to use when \code{method}
is \code{"covariance"}. The default is 30.
\item \code{parallel}: If \code{TRUE}, parallel processing via the
parallel package is automatically used. This can significantly speed
up computations even when \code{method = "all"} (i.e., no big data
approximation is used), as predictions
are spread out over multiple cores. The default is \code{FALSE}.
\item \code{ncores}: If \code{parallel = TRUE}, the number of cores to
parallelize over. The default is the number of available cores on your machine.
}
When \code{local} is a list, at least one list element must be provided to
initialize default arguments for the other list elements.
If \code{local} is \code{TRUE}, defaults for \code{local} are chosen such
that \code{local} is transformed into
\code{list(size = 30, method = "covariance", parallel = FALSE)}.}

\item{grid}{An explicit grid of parameter values by which to evaluate fit. The
names of \code{grid} must contain all the names returned by \code{ssn_decorrelate_grid(formula, data, ...)}.}

\item{dense_grid}{A logical
which controls the density of the constructed grid to be evaluated. If
\code{dense_grid} is \code{TRUE}, a denser grid is used. If \code{dense_grid}
is \code{FALSE}, a sparser grid is used. By default, \code{dense_grid} is \code{FALSE}.}

\item{...}{Other arguments to the functions called by \code{algorithm}.}

\item{x}{A fitted model from \code{\link[=ssn_decorrelate]{ssn_decorrelate()}}.}
}
\value{
A list with many elements that store information about the fitted model object:
\itemize{
\item \code{algorithm}: The machine learning algorithm used.
\item \code{decorrelate_data}: The output of \code{\link[=ssn_decorrelate_data]{ssn_decorrelate_data()}} applied to \code{data}.
\item \code{fit}: The fitted machine learning model object applied to the decorrelated data.
\item \code{grid}: If used, the grid of spatial decorrelation parameters evaluated and their corresponding
metrics when applied to the test data.
\item \code{newdata}: The rows of \code{data} that have \code{NA} response values and are stored as prediction data.
\item \code{training}: If used, the observations assigned to each training and test data set.
\item \code{test}: If used, a list with the lowest (absolute) mean bias (bias), mean-squared-prediction error (MSPE),
root-mean-squared-prediction error (RMSPE), and predictive R-squared (cor2).
}
}
\description{
Apply the spatial decorrelation transformation for SSN data,
allowing for random effects, anisotropy, partition factors, and big data methods.
}
\details{
The spatial decorrelation transformation is a preprocessing transformation
that reduces the impacts of spatial dependence (i.e., covariance, correlation)
on machine learning models. Predictions are made on the
decorrelated scale and then recorrelated to account for spatial dependence.
See Heaton et al., 2025 for details.

\code{tailup_type} Details: Let \eqn{D} be a matrix of hydrologic distances,
\eqn{W} be a diagonal matrix of weights from \code{additive}, \eqn{r = D / range},
and \eqn{I} be
an identity matrix. Then parametric forms for flow-connected
elements of \eqn{R(zu)} are given below:
\itemize{
\item linear: \eqn{(1 - r) * (r <= 1) * W}
\item spherical: \eqn{(1 - 1.5r + 0.5r^3) * (r <= 1) * W}
\item exponential: \eqn{exp(-r) * W}
\item mariah: \eqn{log(90r + 1) / 90r * (D > 0) + 1 * (D = 0) * W}
\item epa: \eqn{(D - range)^2 * F * (r <= 1) * W / 16range^5}
\item gaussian: \eqn{2 exp(-r^2) * (1 - pnorm(r * 2^{1/2})) * W}
\item none: \eqn{I} * W
}

Details describing the \code{F} matrix in the \code{epa} covariance are given in Garreta et al. (2010).
Flow-unconnected elements of \eqn{R(zu)} are assumed uncorrelated.
Observations on different networks are also assumed uncorrelated.

\code{taildown_type} Details: Let \eqn{D} be a matrix of hydrologic distances,
\eqn{r = D / range},
and \eqn{I} be an identity matrix. Then parametric forms for flow-connected
elements of \eqn{R(zd)} are given below:
\itemize{
\item linear: \eqn{(1 - r) * (r <= 1)}
\item spherical: \eqn{(1 - 1.5r + 0.5r^3) * (r <= 1)}
\item exponential: \eqn{exp(-r)}
\item mariah: \eqn{log(90r + 1) / 90r * (D > 0) + 1 * (D = 0)}
\item epa: \eqn{(D - range)^2 * F1 * (r <= 1) / 16range^5}
\item gaussian: \eqn{0}
\item none: \eqn{I}
}

Now let \eqn{A} be a matrix that contains the shorter of the two distances
between two sites and the common downstream junction, \eqn{r1 = A / range},
\eqn{B} be a matrix that contains the longer of the two distances between two sites and the
common downstream junction, \eqn{r2 = B / range},  and \eqn{I} be an identity matrix.
Then parametric forms for flow-unconnected elements of \eqn{R(zd)} are given below:
\itemize{
\item linear: \eqn{(1 - r2) * (r2 <= 1)}
\item spherical: \eqn{(1 - 1.5r1 + 0.5r2) * (1 - r2)^2 * (r2 <= 1)}
\item exponential: \eqn{0}
\item mariah: \eqn{(log(90r1 + 1) - log(90r2 + 1)) / (90r1 - 90r2) * (A =/ B) + (1 / (90r1 + 1)) * (A = B)}
\item epa: \eqn{(B - range)^2 * F2 * (r2 <= 1) / 16range^5}
\item gaussian: \eqn{2 exp(-(B - A) / range) * (1 - pnorm(r * 2^{1/2})) * W}
\item none: \eqn{I}
}

Details describing the \code{F1} and \code{F2} matrices in the \code{epa}
covariance are given in Garreta et al. (2010).
Observations on different networks are assumed uncorrelated.

\code{euclid_type} Details: Let \eqn{D} be a matrix of Euclidean distances,
\eqn{r = D / range}, and \eqn{I} be an identity matrix. Then parametric
forms for elements of \eqn{R(ze)} are given below:
\itemize{
\item exponential: \eqn{exp(- r )}
\item spherical: \eqn{(1 - 1.5r + 0.5r^3) * (r <= 1)}
\item gaussian: \eqn{exp(- r^2 )}
\item cubic: \eqn{(1 - 7r^2 + 8.75r^3 - 3.5r^5 + 0.75r^7) * (r <= 1)}
\item pentaspherical: \eqn{(1 - 1.875r + 1.25r^3 - 0.375r^5) * (r <= 1)}
\item circular: \eqn{1 - (2 / \pi) * (r * sqrt(1 - r^2) + \arcsin(r))} for \eqn{0 \le r \le 1}, and zero for \eqn{r > 1}
\item wave: \eqn{sin(r) * (h > 0) / r + (h = 0)}
\item jbessel: \eqn{Bj(h * range)}, Bj is Bessel-J function
\item gravity: \eqn{(1 + r^2)^{-0.5}}
\item rquad: \eqn{(1 + r^2)^{-1}}
\item magnetic: \eqn{(1 + r^2)^{-1.5}}
\item matern: \eqn{2^{1-extra} eta^{extra} K_{extra}(eta) / Gamma(extra)}, where \eqn{eta = \sqrt{2 extra} D / range}
\item cauchy: \eqn{(1 + r^2)^{-extra}}
\item pexponential: \eqn{exp(-D^{extra} / range)}
\item none: \eqn{I}
}
The powered-exponential range has units of distance raised to \code{extra}.

\code{nugget_type} Details: Let \eqn{I} be an identity matrix and \eqn{0}
be the zero matrix. Then parametric
forms for elements the nugget variance are given below:
\itemize{
\item nugget: \eqn{I}
\item none: \eqn{0}
}
In short, the nugget effect is modeled when \code{nugget_type} is \code{"nugget"}
and omitted when \code{nugget_type} is \code{"none"}.

\code{estmethod} Details: The various estimation methods are
\itemize{
\item \code{reml}: Maximize the restricted log-likelihood.
\item \code{ml}: Maximize the log-likelihood.
}

\code{anisotropy} Details: By default, all Euclidean covariance parameters except \code{rotate}
and \code{scale} as well as all random effect variance parameters
are assumed unknown, requiring estimation. If either \code{rotate} or \code{scale}
are given initial values other than 0 and 1 (respectively) or are assumed unknown
in \code{\link[=euclid_initial]{euclid_initial()}}, \code{anisotropy} is implicitly set to \code{TRUE}.
(Geometric) Anisotropy is modeled by transforming a Euclidean covariance function that
decays differently in different directions to one that decays equally in all
directions via rotation and scaling of the original Euclidean coordinates. The rotation is
controlled by the \code{rotate} parameter in \eqn{[0, \pi]} radians. The scaling
is controlled by the \code{scale} parameter in \eqn{[0, 1]}. The anisotropy
correction involves first a rotation of the coordinates clockwise by \code{rotate} and then a
scaling of the coordinates' minor axis by the reciprocal of \code{scale}. The Euclidean
covariance is then computed using these transformed coordinates.

\code{random} Details: If random effects are used, the model
can be written as \eqn{y = X \beta + Z1u1 + ... Zjuj + zu + zd + ze + n},
where each Z is a random effects design matrix and each u is a random effect.

\code{partition_factor} Details: The partition factor can be represented in matrix form as \eqn{P}, where
elements of \eqn{P} equal one for observations in the same level of the partition
factor and zero otherwise. The covariance matrix involving only the
spatial and random effects components is then multiplied element-wise
(Hadamard product) by \eqn{P}, yielding the final covariance matrix.

\code{training} Details: When \code{replicate} or the number of cross validation folds is at
least two, there are separate grids evaluated for each replication (or fold). Statistics in each grid are
averaged across replications (or folds) to determine a final grid ranked by \code{statistic}.

\code{local} Details: The big data approximation works by leveraging the
conditional nature of the spatial decorrelation transformation via the
Vecchia approximation. The Vecchia approximation enables efficient computation
of the conditional distribution by considering only the \code{size} most relevant
observations in the ordering (rather than using all the observations).

Observations with \code{NA} response values are removed for model
fitting, but their values can be predicted afterwards by running
\code{predict(object)}.
}
\examples{
\donttest{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path,
  predpts = "pred1km", overwrite = TRUE
)
ssn_create_distmat(mf04p, predpts = "pred1km", overwrite = TRUE)
fit <- ssn_decorrelate(
  Summer_mn ~ ELEV_DEM, mf04p, tailup_type = "exponential",
  taildown_type = "exponential", euclid_type = "exponential",
  additive = "afvArea", dense_grid = FALSE
)
tidy(fit$grid)
head(predict(fit, "pred1km"))
}
}
\references{
Matthew J. Heaton, Andrew Millane, and Jake S. Rhodes. 2025. A Scalable
Spatial Decorrelation Preprocessing Approach for Machine and Deep Learning.
\emph{Journal of Data Science}. 1-15, DOI 10.6339/25-JDS1210
}
