% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate_ssn.R
\name{ssn_decorrelate_newdata}
\alias{ssn_decorrelate_newdata}
\title{Apply the Spatial Decorrelation Transformation to a Newdata Object for Prediction}
\usage{
ssn_decorrelate_newdata(object, newdata, local, ...)
}
\arguments{
\item{object}{An \code{\link[=ssn_decorrelate_data]{ssn_decorrelate_data()}} object.}

\item{newdata}{A character vector that indicates the name of the prediction data set
for which predictions are desired (accessible via \code{object$ssn.object$preds}).
Note that the prediction data must be in the original SSN object used to fit the model.
If \code{newdata} is omitted, predictions
for all prediction data sets are returned. Note that the name \code{".missing"}
indicates the prediction data set that contains the missing observations in the data used
to fit the model.}

\item{local}{An optional logical or list controlling the big data approximation.
If omitted, \code{local} is set
to \code{TRUE} or \code{FALSE} based on the sample size (the number of
non-missing observations in \code{data}) -- if the sample size exceeds 5,000,
\code{local} is set to \code{TRUE}. Otherwise it is set to \code{FALSE}.
If \code{local} is \code{FALSE}, no big data approximation
is implemented. If a list is provided, the following arguments detail the big
data approximation:
\itemize{
\item \code{method}: The big data approximation method. If \code{method = "all"},
all observations are used and \code{size} is ignored.
If \code{method = "covariance"}, the \code{size} data observations
with the highest covariance with the observation requiring prediction are used.
The default is \code{"covariance"}.
\item \code{size}: The number of data observations to use when \code{method}
is \code{"distance"} or \code{"covariance"}. The default is 30.
\item \code{parallel}: If \code{TRUE}, parallel processing via the
parallel package is automatically used. This can significantly speed
up computations even when \code{method = "all"} (i.e., no big data
approximation is used), as predictions
are spread out over multiple cores. The default is \code{FALSE}.
\item \code{ncores}: If \code{parallel = TRUE}, the number of cores to
parallelize over. The default is the number of available cores on your machine.
}
When \code{local} is a list, at least one list element must be provided to
initialize default arguments for the other list elements.
If \code{local} is \code{TRUE}, defaults for \code{local} are chosen such
that \code{local} is transformed into
\code{list(size = 30, method = "covariance", parallel = FALSE)}.}

\item{...}{Other arguments.}
}
\value{
A list with many elements that store information about
the fitted model object. Importantly, the list contains the following element:
\itemize{
\item \code{X_newdata}: The original fixed effects design matrix (of explanatory variables) for the prediction data.
\item \code{tX_newdata}: The spatially decorrelated fixed effects design matrix for the prediction data.
}
}
\description{
Apply the spatial decorrelation transformation to a newdata object.
This object contains explanatory variables that are transformed for prediction
according to some spatial decorrelation transformation.
}
\examples{
\donttest{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path,
  predpts = "CapeHorn", overwrite = TRUE
)
ssn_create_distmat(mf04p, predpts = "CapeHorn", overwrite = TRUE)
transformed <- ssn_decorrelate_data(
  Summer_mn ~ ELEV_DEM, mf04p, additive = "afvArea",
  tailup_params = tailup_params("exponential", de = 1, range = 10000),
  nugget_params = nugget_params("nugget", nugget = 0.1)
)
new_transformed <- ssn_decorrelate_newdata(transformed, "CapeHorn")
head(new_transformed$tX_newdata)
}
}
