% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_import.R
\name{ssn_import}
\alias{ssn_import}
\title{Import \code{SSN} object}
\usage{
ssn_import(
  path,
  include_obs = TRUE,
  predpts = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Filepath to the .ssn directory. See details.}

\item{include_obs}{default = \code{TRUE}. Logical indicating
whether observed sites should be included in the SSN object.}

\item{predpts}{Vector of prediction site dataset names found within
the .ssn folder. See details.}

\item{overwrite}{default = \code{FALSE}. If \code{TRUE}, overwrite
existing binaryID.db files and netgeom column(s) if they exist in the
edges, observed sites (if \code{include_obs = TRUE}), and
prediction site datasets (if they exist).}

\item{verbose}{default = \code{TRUE}. If \code{FALSE}, warning messages will not be printed to the console.}
}
\value{
\code{ssn_import} returns an object of class SSN, which is a list
with four elements containing:
\itemize{
\item \code{edges}: An \code{sf data.frame} containing the stream network,
with an additional 'netgeom' column.
\item \code{obs}: An sf data.frame containing observed site locations,
with an additional 'netgeom' column. NA if \code{include_obs =
    FALSE}.
\item \code{preds}: A list of sf data.frames containing prediction
site locations. An empty list is returned if \code{predpts} is not provided.
\item path: The local filepath for the .ssn directory associated with the \code{SSN}
object.
}
}
\description{
This function reads spatial data from a .ssn folder
and creates an \code{SSN} object.
}
\details{
The \command{ssn_import} function imports spatial data (shapefile or GeoPackage format)
from a .ssn folder generated using the
\code{SSNbler} package function \code{SSNbler::lsn_to_ssn}. The .ssn folder contains all of the spatial, topological and
attribute data needed to fit a spatial statistical stream network
model to streams data.  This includes:
\itemize{
\item An edges dataset with LINESTRING geometry representing the stream network.
\item A sites dataset with POINT geometry where observed data were collected on
the stream network.
\item Prediction sites dataset(s) representing
locations where predictions will be made.
\item netID.dat file(s) for each distinct network, which stores the topological
relationships of the line features in edges.
}

A more detailed description of the .ssn directory and its
contents is provided in Peterson and Ver Hoef (2014).

The \command{ssn_import} imports the edges, observed sites (optional), and
prediction sites (optional) as \code{sf data.frame} objects. A new column named 'netgeom'
is created to store important data representing
topological relationships in a spatial stream network
model. These data are stored in character format, which is less
likely to be inadvertently changed by users. See
\code{\link[SSN2]{create_netgeom}} for a more detailed description of
the format and contents of 'netgeom'.

The information contained in the netID text files is imported
into an SQLite database, binaryID.db, which is stored in the .ssn
directory. This information is used internally by
\code{\link[SSN2]{ssn_create_distmat}},
\code{\link[SSN2]{ssn_lm}} and
\code{\link[SSN2]{ssn_glm}} to calculate the data necessary
to fit a spatial statistical model to stream network data. If
\code{overwrite = TRUE} (\code{overwrite = FALSE} is the default) and a binaryID.db
file already exists within the .ssn directory, it will be
overwritten when the \code{SSN} object is created. If a 'netgeom' column exists in any of the input datasets (e.g. edges, observed sites, prediction sites) and \code{overwrite = TRUE}, it will be overwritten.

At a minimum, an \code{SSN} object must always contain streams,
which are referred to as edges. The \code{SSN} object would also
typically contain a set of observed sites, where measurements
have been collected. Only one observed dataset is permitted in an
\code{SSN} object. When \code{include_obs=FALSE}, an \code{SSN}
object is created without observations. This option provides
flexibility for users who would like to simulate data on a set of
artificial sites on an existing stream network. Note that
observation sites must be included in the \code{SSN} object in
order to fit models using \command{ssn_lm} or
\command{ssn_glm}. The \code{SSN} object may contain multiple
sets of prediction points (or none), which are stored as separate
datasets in the .ssn directory. If \code{predpts} is a named
vector, the names of the \code{preds} list in the \code{SSN}
object correspond to the vector names. Otherwise, they are set to
the basename of the prediction site dataset file(s) specified in
\code{predpts}. The \code{\link[SSN2]{ssn_import_predpts}}
function allows users to import additional sets of prediction
sites to an existing \code{SSN} object.
}
\examples{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")

## Import SSN object with no prediction sites
mf04 <- ssn_import(temp_path,
  overwrite = TRUE
)

## Import SSN object with 3 sets of prediction sites
mf04p <- ssn_import(temp_path,
  predpts = c(
    "pred1km",
    "CapeHorn"
  ),
  overwrite = TRUE
)

}
\references{
Peterson, E., and Ver Hoef, J.M. (2014) STARS: An
ArcGIS toolset used to calculate the spatial information needed
to fit spatial statistical stream network models to stream
network data. \emph{Journal of Statistical Software}
\bold{56(2)}, 1--17.
}
