% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_lmRF.R
\name{ssn_lmRF}
\alias{ssn_lmRF}
\title{Fit a random forest stream-network residual model}
\usage{
ssn_lmRF(formula, ssn.object, ...)
}
\arguments{
\item{formula}{A two-sided linear formula describing the fixed effect structure
of the model, with the response to the left of the \code{~} operator and
the terms on the right, separated by \code{+} operators.}

\item{ssn.object}{A spatial stream network object with class \code{SSN}.}

\item{...}{Additional named arguments to \code{\link[ranger:ranger]{ranger::ranger()}} or \code{\link[=ssn_lm]{ssn_lm()}}.}
}
\value{
A list with several elements to be used with \code{predict()}. These
elements include the function call (named \code{call}), the random forest object
fit to the mean (named \code{ranger}),
the spatial stream network model object fit to the residuals
(named \code{ssn_lm}), and an object can contain data for
locations at which to predict (called \code{newdata}). The \code{newdata}
object contains the set of
observations in \code{data} whose response variable is \code{NA}.
}
\description{
Fit random forest spatial residual models for stream network data
using random forest to fit the mean and a spatial linear model to fit the residuals.
The spatial linear model fit to the residuals can incorporate variety of estimation
methods, allowing for random effects, anisotropy, partition factors, and big data methods.
}
\details{
The random forest residual spatial linear model is described by
Fox et al. (2020). A random forest model is fit to the mean portion of the
model specified by \code{formula} using \code{ranger::ranger()}. Residuals
are computed and used as the response variable in an intercept-only spatial
linear model fit using \code{\link[=ssn_lm]{ssn_lm()}}. This model object is intended for use with
\code{predict()} to perform prediction, also called random forest
regression Kriging.
}
\note{
This function does not perform any internal scaling. If optimization is not
stable due to extremely large variances, scale relevant variables
so they have variance 1 before optimization.
}
\examples{
if (requireNamespace("ranger", quietly = TRUE)) {
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf <- ssn_import(temp_path,
  predpts = "CapeHorn", overwrite = TRUE
)
ssn_create_distmat(mf, predpts = "CapeHorn", overwrite = TRUE)
rf_fit <- ssn_lmRF(
  Summer_mn ~ ELEV_DEM, mf, tailup_type = "exponential",
  additive = "afvArea", num.trees = 50, num.threads = 1, seed = 1
)
print(rf_fit)
head(predict(rf_fit, "CapeHorn"))
}
}
\references{
Fox, E.W., Ver Hoef, J.M., & Olsen, A.R. (2020). Comparing spatial
regression to random forests for large environmental data sets.
\emph{PLoS ONE}, 15(3), e0229509.
}
