% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_names.R
\name{ssn_names}
\alias{ssn_names}
\title{Return names of data in an SSN object}
\usage{
ssn_names(ssn.object)
}
\arguments{
\item{ssn.object}{An \code{SSN} object.}
}
\value{
Print variable names to console
}
\description{
Extract and print names from the \code{edges}, \code{sites}
and \code{preds} elements of an SSN object.
}
