% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate_ssn.R
\name{ssn_recorrelate_newdata}
\alias{ssn_recorrelate_newdata}
\title{Recorrelate Machine Learning Predictions}
\usage{
ssn_recorrelate_newdata(object, ty_newdata)
}
\arguments{
\item{object}{An \code{\link[=ssn_decorrelate_newdata]{ssn_decorrelate_newdata()}} object.}

\item{ty_newdata}{Predictions from the machine learning model trained on the
spatially decorrelated data and applied to spatially decorrelated newdata.}
}
\value{
A vector of predictions
}
\description{
Recorrelate machine learning predictions according to a
spatial decorrelation transformation.
}
\examples{
\donttest{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path,
  predpts = "CapeHorn", overwrite = TRUE
)
ssn_create_distmat(mf04p, predpts = "CapeHorn", overwrite = TRUE)
transformed <- ssn_decorrelate_data(
  Summer_mn ~ ELEV_DEM, mf04p, additive = "afvArea",
  tailup_params = tailup_params("exponential", de = 1, range = 10000),
  nugget_params = nugget_params("nugget", nugget = 0.1)
)
learner <- lm.fit(transformed$tX, transformed$ty)
new_transformed <- ssn_decorrelate_newdata(transformed, "CapeHorn")
predictions <- ssn_recorrelate_newdata(
  new_transformed, as.numeric(new_transformed$tX_newdata \%*\% learner$coefficients)
)
head(predictions)
}
}
