% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSN_to_SSN2.R
\name{SSN_to_SSN2}
\alias{SSN_to_SSN2}
\title{Convert object from \code{SpatialStreamNetwork} class to \code{SSN} class}
\usage{
SSN_to_SSN2(object)
}
\arguments{
\item{object}{A SpatialStreamNetwork object}
}
\value{
An S3 \code{SSN} class object.
}
\description{
Convert an S4 \code{SpatialStreamNetwork} object
created in the SSN package to an S3 \code{SSN} object used in the
SSN2 package.
}
\details{
\command{SSN_to_SSN2()} has been made available to help users
migrate from the \code{SSN} package to the updated \code{SSN2}
package. It is used to convert existing S4 \code{SpatialStreamNetwork}
objects \emph{stored in saved workspaces} to the S3 \code{SSN} class object
used in the \code{SSN2} package. Note that \code{\link[SSN2]{ssn_import}} is
used to create an S3 \code{SSN} object from data stored locally in a .ssn
directory.
}
