% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_update_path.R
\name{ssn_update_path}
\alias{ssn_update_path}
\title{Update path in an SSN object}
\usage{
ssn_update_path(x, path, verbose = FALSE)
}
\arguments{
\item{x}{An SSN, ssn_lm or ssn_glm object.}

\item{path}{Filepath to the .ssn folder associated with the SSN
object.}

\item{verbose}{A logical that indicates if the new path should be printed
to the console.}
}
\value{
An SSN object with a new path list element.
}
\description{
Update the local path in an existing SSN object based
on a user-defined file.
}
\details{
At times, it may be necessary to move a .ssn directory,
which is linked to an SSN object in an R workspace. If the .ssn
directory is moved, the path must be updated before using the
\code{ssn_lm()}/\code{ssn_glm()} functions and other functions that read/write
to the .ssn directory. The \command{ssn_update_path} is a helper
function that serves this purpose.
}
\examples{
## Use mf04p SSN object provided in SSN2
data(mf04p)

## For examples only, make sure mf04p has the correct path
## If you use ssn_import(), the path will be correct
newpath <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_update_path(mf04p, newpath)
}
