% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SSN.R
\name{summary.SSN}
\alias{summary.SSN}
\title{Summarize an SSN object}
\usage{
\method{summary}{SSN}(object, ...)
}
\arguments{
\item{object}{An SSN object.}

\item{...}{Other arguments. Not used (needed for generic consistency).}
}
\value{
A list with several fitted model quantities used to create
informative summaries when printing.
}
\description{
Summarize data found in an SSN object.
}
\details{
\code{summary.SSN()} creates a summary of an SSN object
intended to be printed using \code{print()}. This summary
contains information about the number of observed and prediction
locations, as well as the column names found in their respective
sf data.frames.
}
