% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amongSitesBigDistMat.R
\name{amongSitesBigDistMat}
\alias{amongSitesBigDistMat}
\title{Helper function to determining distance matrices among sites}
\usage{
amongSitesBigDistMat(
  ssn,
  pids,
  net.num,
  name = "obs",
  bin.table,
  workspace.name
)
}
\arguments{
\item{ssn}{An SSN object.}

\item{pids}{A list of pid values for prediction sites}

\item{net.num}{Network number (netID) in character format.}

\item{name}{The network name (obs or prediction name)}

\item{bin.table}{A binaryID table for the network.}

\item{workspace.name}{Name of new distance matrix file}
}
\value{
A distance matrix
}
\description{
Helper function to determining distance matrices among sites
}
