% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SSN2}
\alias{print.SSN2}
\alias{print.ssn_lm}
\alias{print.ssn_glm}
\alias{print.summary.ssn_lm}
\alias{print.summary.ssn_glm}
\alias{print.anova.ssn_lm}
\alias{print.anova.ssn_glm}
\title{Print values}
\usage{
\method{print}{ssn_lm}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{ssn_glm}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{summary.ssn_lm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{summary.ssn_glm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{anova.ssn_lm}(
  x,
  digits = max(getOption("digits") - 2L, 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{anova.ssn_glm}(
  x,
  digits = max(getOption("digits") - 2L, 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A fitted model object from \code{\link[=ssn_lm]{ssn_lm()}}, a fitted model object from \code{\link[=ssn_glm]{ssn_glm()}},
or output from \code{summary(x)} or or \code{anova(x)}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Other arguments passed to or from other methods.}

\item{signif.stars}{Logical. If \code{TRUE}, significance stars are printed for each coefficient}
}
\value{
Printed fitted model objects and summaries with formatting.
}
\description{
Print fitted model objects and summaries.
}
