% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSN2.R
\docType{package}
\name{SSN2-package}
\alias{SSN2}
\alias{SSN2-package}
\title{SSN2: Spatial Modeling on Stream Networks}
\description{
Spatial statistical modeling and prediction for data on stream networks, including models based on in-stream distance (Ver Hoef, J.M. and Peterson, E.E., 2010. \doi{10.1198/jasa.2009.ap08248}.) Models are created using moving average constructions. Spatial linear models, including explanatory variables, can be fit with (restricted) maximum likelihood. Mapping and other graphical functions are included.
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Authors:
\itemize{
  \item Jay M. Ver Hoef \email{jay.verhoef@noaa.gov}
  \item Erin Peterson
  \item Alan Pearse
  \item Dan Isaak
}

}
\keyword{internal}
