% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate_ssn.R
\name{ssn_decorrelate_grid}
\alias{ssn_decorrelate_grid}
\alias{print.ssn_decorrelate_grid}
\alias{tidy.ssn_decorrelate_grid}
\title{Create a Spatial Decorrelation Transformation Grid}
\usage{
ssn_decorrelate_grid(
  formula,
  ssn.object,
  tailup_type = "none",
  taildown_type = "none",
  euclid_type = "none",
  nugget_type = "nugget",
  tailup_params,
  taildown_params,
  euclid_params,
  nugget_params,
  additive,
  anisotropy = FALSE,
  random,
  randcov_params,
  dense_grid
)

\method{print}{ssn_decorrelate_grid}(x, ...)

\method{tidy}{ssn_decorrelate_grid}(x, sort_by, decreasing, ...)
}
\arguments{
\item{formula}{A two-sided linear formula describing the fixed effect structure
of the model, with the response to the left of the \code{~} operator and
the terms on the right, separated by \code{+} operators. \code{.} on the
right-hand side represents every variable in \code{data} except the geometry column.}

\item{ssn.object}{A spatial stream network object with class \code{SSN}.}

\item{tailup_type, taildown_type, euclid_type, nugget_type}{Covariance types
specified as in \code{\link[=ssn_lm]{ssn_lm()}}.}

\item{tailup_params, taildown_params, euclid_params, nugget_params}{Known
covariance parameter objects to be used in the spatial decorrelation
transformation. See \code{\link[=tailup_params]{tailup_params()}}, \code{\link[=taildown_params]{taildown_params()}},
\code{\link[=euclid_params]{euclid_params()}}, and \code{\link[=nugget_params]{nugget_params()}}.}

\item{additive}{Additive-function variable for the tail-up covariance, as in \code{\link[=ssn_lm]{ssn_lm()}}.}

\item{anisotropy}{A logical indicating whether Euclidean (geometric) anisotropy should
be modeled. Not required if \code{euclid_params} is provided with a nonzero
\code{rotate} or a \code{scale} less than one. When \code{anisotropy} is
\code{TRUE}, computational times can significantly increase. The default
is \code{FALSE}.}

\item{random}{A one-sided linear formula describing the random effect structure
of the model. Terms are specified to the right of the \code{~ operator}.
Each term has the structure \code{x1 + ... + xn | g1/.../gm}, where \code{x1 + ... + xn}
specifies the model for the random effects and \code{g1/.../gm} is the grouping
structure. Separate terms are separated by \code{+} and must generally
be wrapped in parentheses. Random intercepts are added to each model
implicitly when at least one other variable is defined.
If a random intercept is not desired, this must be explicitly
defined (e.g., \code{x1 + ... + xn - 1 | g1/.../gm}). If only a random intercept
is desired for a grouping structure, the random intercept must be specified
as \code{1 | g1/.../gm}. Note that \code{g1/.../gm} is shorthand for \code{(1 | g1/.../gm)}.
If only random intercepts are desired and the shorthand notation is used,
parentheses can be omitted.}

\item{randcov_params}{An optional random effect covariance object
to be used in the spatial decorrelation transformation. See \code{\link[spmodel:randcov_params]{spmodel::randcov_params()}}.}

\item{dense_grid}{A logical
which controls the density of the constructed grid to be evaluated. If
\code{dense_grid} is \code{TRUE}, a denser grid is used. If \code{dense_grid}
is \code{FALSE}, a sparser grid is used. By default, \code{dense_grid}
is \code{FALSE} when the sample size is greater than 5,000 and \code{TRUE}
otherwise.}

\item{x}{An object from \code{object$grid}.}

\item{...}{Other arguments to the functions called by \code{algorithm}.}

\item{sort_by}{Sort by a specific row in \code{x}. Fit statistics are
\code{"bias"}, \code{"MSPE"}, \code{"RMSPE"}, and \code{"cor2"}.
The default is \code{"MSPE"}.}

\item{decreasing}{Whether \code{sort_by} should sort in decreasing order?
If \code{sort_by = "cor2"}, the default is \code{TRUE}; otherwise it is
\code{FALSE}.}
}
\value{
A grid of spatial decorrelation parameters stored as a \code{data.frame}.
}
\description{
Create a spatial decorrelation transformation grid of initial parameters to be
evaluated via a grid search.
}
\examples{
# Copy the mf04p .ssn data to a local directory and read it into R
# When modeling with your .ssn object, you will load it using the relevant
# path to the .ssn data on your machine
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(temp_path, overwrite = TRUE)
grid <- ssn_decorrelate_grid(
  Summer_mn ~ ELEV_DEM, mf04p, tailup_type = "exponential",
  additive = "afvArea", dense_grid = FALSE
)
tidy(grid)
# Supply edited rows as grid = custom_grid to ssn_decorrelate().
custom_grid <- grid[1:2, ]
custom_grid$tailup_range <- c(5000, 10000)
}
