% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{ssn_initial}
\alias{ssn_initial}
\alias{tailup_initial}
\alias{taildown_initial}
\alias{euclid_initial}
\alias{nugget_initial}
\title{Create a covariance parameter initial object}
\usage{
tailup_initial(tailup_type, de, range, known)

taildown_initial(taildown_type, de, range, known)

euclid_initial(euclid_type, de, range, rotate, scale, known)

nugget_initial(nugget_type, nugget, known)
}
\arguments{
\item{tailup_type}{The tailup covariance function type. Available options
include \code{"linear"}, \code{"spherical"}, \code{"exponential"},
\code{"mariah"}, \code{"epa"}, and \code{"none"}.}

\item{de}{The spatially dependent (correlated) random error variance. Commonly referred to as
a partial sill.}

\item{range}{The correlation parameter.}

\item{known}{A character vector indicating which covariance parameters are to be
assumed known. The value \code{"given"} is shorthand for assuming all
covariance parameters given to \code{*_initial()} are assumed known.}

\item{taildown_type}{The taildown covariance function type. Available options
include \code{"linear"}, \code{"spherical"}, \code{"exponential"},
\code{"mariah"}, \code{"epa"}, and \code{"none"}.}

\item{euclid_type}{The euclidean covariance function type. Available options
include \code{"spherical"}, \code{"exponential"}, \code{"gaussian"},
\code{"cosine"}, \code{"cubic"}, \code{"pentaspherical"}, \code{"wave"},
\code{"jbessel"}, \code{"gravity"}, \code{"rquad"}, \code{"magnetic"}, and
\code{"none"}.}

\item{rotate}{Anisotropy rotation parameter (from 0 to \eqn{\pi} radians) for
the euclidean portion of the covariance. A value of 0 (the default) implies no rotation.}

\item{scale}{Anisotropy scale parameter (from 0 to 1) for
the euclidean portion of the covariance. A value of 1 (the default) implies no scaling.}

\item{nugget_type}{The nugget covariance function type. Available options
include \code{"nugget"} or \code{"none"}.}

\item{nugget}{The spatially independent (not correlated) random error variance. Commonly referred to as
a nugget.}
}
\description{
Create a covariance parameter initial object that specifies
initial and/or known values to use while estimating specific covariance parameters
with \code{\link[=ssn_lm]{ssn_lm()}} or \code{\link[=ssn_glm]{ssn_glm()}}.  See \code{\link[spmodel:randcov_initial]{spmodel::randcov_initial()}} for documentation regarding
random effect covariance parameter initial objects.
}
\examples{
tailup_initial("exponential", de = 1, range = 20, known = "range")
tailup_initial("exponential", de = 1, range = 20, known = "given")
euclid_initial("spherical", de = 2, range = 4, scale = 0.8, known = c("range", "scale"))
dispersion_initial("nbinomial", dispersion = 5)

}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach
to geostatistical modeling in stream networks. \emph{Ecology} \bold{91(3)},
644--651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial
statistical models of stream networks (with discussion).
\emph{Journal of the American Statistical Association} \bold{105}, 6--18.
DOI: 10.1198/jasa.2009.ap08248.  Rejoinder pgs. 22--24.
}
